backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
face-recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy>=1.21.0,<2.0.0
scikit-learn==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0